+++ ThunderStore/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+- Coins no longer spawn outside (causes issues)
+- Coins should despawn properly now
+
 ## 1.0.1
 - Removed LethalLib as a dependency as it is not needed
 - Increased the collider for the coin to make it easier to collect